chore: migrate images to separate API
