+++ apps/api/src/lib/api/getProfileMeta.ts
@@ -1,15 +1,10 @@
-import { readFileSync } from 'node:fs';
-import path from 'node:path';
-
 import { gql } from '@apollo/client';
 import generateMeta from '@lib/generateMeta';
-import { Resvg } from '@resvg/resvg-js';
+import getIPFSLink from '@lib/getIPFSLink';
 import type { MediaSet, NftImage, Profile } from 'lens';
 import type { NextApiRequest, NextApiResponse } from 'next';
-import satori from 'satori';
 import client from 'src/apollo';
 
-import { markUp } from './getMarkup';
 const PROFILE_QUERY = gql`
   query Profile($request: SingleProfileQueryRequest!) {
     profile(request: $request) {
@@ -43,58 +38,22 @@ const getProfileMeta = async (req: NextApiRequest, res: NextApiResponse, handle:
 
     if (data?.profile) {
       const profile: Profile & { picture: MediaSet & NftImage } = data?.profile;
-
-      const fontNormalPath = path.resolve('./fonts', 'CircularXXSub-Book.woff');
-      const fontMediumPath = path.resolve('./fonts', 'CircularXXSub-Medium.woff');
-      const fontBoldPath = path.resolve('./fonts', 'CircularXXSub-Bold.woff');
-
-      const svg = await satori(markUp(profile), {
-        width: 1200,
-        height: 600,
-        fonts: [
-          {
-            name: 'CircularXX Normal',
-            data: readFileSync(fontNormalPath),
-            weight: 400,
-            style: 'normal'
-          },
-          {
-            name: 'CircularXX Bold',
-            data: readFileSync(fontBoldPath),
-            weight: 400,
-            style: 'normal'
-          },
-          {
-            name: 'CircularXX Medium',
-            data: readFileSync(fontMediumPath),
-            weight: 400,
-            style: 'normal'
-          }
-        ]
-      });
-      const resvg = new Resvg(svg, {
-        background: 'rgba(238, 235, 230, .9)',
-        fitTo: {
-          mode: 'width',
-          value: 1200
-        },
-        font: {
-          fontFiles: [
-            path.resolve('./fonts', 'CircularXXSub-Book.woff'),
-            path.resolve('./fonts', 'CircularXXSub-Bold.woff'),
-            path.resolve('./fonts', 'CircularXXSub-Medium.woff')
-          ],
-          loadSystemFonts: false,
-          defaultFontFamily: 'CircularXX Normal'
-        }
-      });
-      const pngData = resvg.render();
-      const pngBuffer = pngData.asPng();
+      const title = profile?.name
+        ? `${profile?.name} (@${profile?.handle}) • Lenster`
+        : `@${profile?.handle} • Lenster`;
+      const description = profile?.bio ?? '';
+      const image = profile
+        ? `https://ik.imagekit.io/lensterimg/tr:n-avatar/${getIPFSLink(
+            profile?.picture?.original?.url ??
+              profile?.picture?.uri ??
+              `https://avatar.tobi.sh/${profile?.ownedBy}_${profile?.handle}.png`
+          )}`
+        : 'https://assets.lenster.xyz/images/og/logo.jpeg';
 
       return res
-        .setHeader('Content-Type', 'image/png')
+        .setHeader('Content-Type', 'text/html')
         .setHeader('Cache-Control', 's-maxage=86400')
-        .send(pngBuffer);
+        .send(generateMeta(title, description, image, btoa(escape(JSON.stringify(profile)))));
     }
   } catch (error) {
     console.error('Error: cannot generate og image', error);
